precision problem fixed

+++ src/model/transformer.py
@@ -121,7 +121,7 @@ class TransformerEncoder(nn.Module):
         """
         x = self.embedding(src_tokens) * self.embed_scale
         x = x + self.positional_embedding(src_tokens)
-        x = x * (1. - src_key_padding_mask.float().unsqueeze(2))
+        x = x * (1. - src_key_padding_mask.type_as(x).unsqueeze(2))
         x = F.dropout(x, p=self.embed_dropout, training=self.training)
 
         x = x.transpose(0, 1)
@@ -192,7 +192,7 @@ class TransformerDecoder(nn.Module):
     def forward(self, encoder_out, tgt_tokens, src_key_padding_mask, tgt_key_padding_mask, tgt_mask, cache=None):
         x = self.embedding(tgt_tokens) * self.embed_scale
         x = x + self.positional_embedding(tgt_tokens)
-        x = x * (1. - tgt_key_padding_mask.float().unsqueeze(2))
+        x = x * (1. - tgt_key_padding_mask.type_as(x).unsqueeze(2))
         x = F.dropout(x, p=self.embed_dropout, training=self.training)
 
         x = x.transpose(0, 1)
